Add Nix Development Environment
Making use of https://devenv.sh, this adds an easy* to use development
environment.
* Provided you have nix installed and direnv configured

This is based largely on the previous running-locally instructions, but
should be a bit more portable.
Due to this, this isn't necessarily the most nix-esque packaging of this
(still making use of venv rather than native nix packaging of python
modules), but this should make it quite low-maintenance compared to a
full repackaging.

This has not (yet) been tested on any platform other than x86_64 linux.