Allow for running without GOOGLE_CREDS
Previously, if GOOGLE_CREDS was set to None (ie: for local development)
the gwallet module would fail to initialise as it couldn't construct the
client.

This fixes this, and replaces the fatal error with an obnoxious warning
that running without google credentials may have unexpected
consequinces. Don't say you weren't warned.